added test setup and api tests

+++ backend/src/Store.Core/ViewModel/AppUserModel.cs
@@ -9,7 +9,7 @@
         public string Role { get; set; }
         public string Password { get; set; }
         public string Token { get; set; }
-        public int? Discount { get; set; }
+        public decimal? Discount { get; set; }
 
 
     }
